Fix intl dropdown not being fixed on navbar
