+++ src/components/Page/Page.js
@@ -2,12 +2,9 @@ import React from 'react'
 import { Link, SiteData, RouteData, Head } from 'react-static'
 import styled from 'styled-components'
 import { AppBar, BaseStyles, PublicUrl, breakpoint } from '@aragon/ui'
-import Dropdown from 'react-bootstrap/Dropdown'
 import { FormattedMessage } from 'react-intl'
 import Navbar from './Navbar'
 import Footer from './Footer.js'
-import en from './assets/en.svg'
-import es from './assets/es.svg'
 const medium = css => breakpoint('medium', css)
 const large = css => breakpoint('large', css)
 
@@ -65,45 +62,6 @@ class Page extends React.Component {
                   </p>
                 </AGPBanner>
                 <Navbar menuItems={items} path={path} color={color} />
-                <Intl>
-                  <Dropdown>
-                    <Dropdown.Toggle id="dropdown-basic">
-                      {localStorage.getItem('locale') == 'es' ? (
-                        <img src={es} />
-                      ) : (
-                        <img src={en} />
-                      )}
-                    </Dropdown.Toggle>
-                    <Dropdown.Menu>
-                      <Dropdown.Item
-                        onClick={() => {
-                          localStorage.setItem('locale', 'en')
-                          if (window.location.href.indexOf('/es/') >= 0) {
-                            let url = window.location.href.split('/es/')
-                            window.location.href = url[0] + '/en/' + url[1]
-                          } else {
-                            window.location.reload(false)
-                          }
-                        }}
-                      >
-                        <img src={en} /> English
-                      </Dropdown.Item>
-                      <Dropdown.Item
-                        onClick={() => {
-                          localStorage.setItem('locale', 'es')
-                          if (window.location.href.indexOf('/en/') >= 0) {
-                            let url = window.location.href.split('/en/')
-                            window.location.href = url[0] + '/es/' + url[1]
-                          } else {
-                            window.location.reload(false)
-                          }
-                        }}
-                      >
-                        <img src={es} /> Español
-                      </Dropdown.Item>
-                    </Dropdown.Menu>
-                  </Dropdown>
-                </Intl>
                 <Content contentColor={contentColor}>{children}</Content>
                 <Footer path={path} />
               </PublicUrl.Provider>
@@ -114,34 +72,6 @@ class Page extends React.Component {
     )
   }
 }
-const Intl = styled.div`
-  position: absolute;
-  top: 10px;
-  top: 74px;
-  ${medium('top: 44px;')}
-  right: 10px;
-  z-index: 9999999;
-  .dropdown-menu {
-    width: 30px;
-    left: auto !important;
-    right: 0px !important;
-    transform: translate3d(-6px, 39px, 0px) !important;
-  }
-  .btn-primary,
-  .btn-primary:hover,
-  .btn-primary:active,
-  .btn-primary:not(:disabled):not(.disabled).active,
-  .btn-primary:not(:disabled):not(.disabled):active,
-  .show > .btn-primary.dropdown-toggle {
-    color: black;
-    display: flex;
-    align-items: center;
-    background-color: transparent;
-    border-color: transparent;
-    outline: none;
-    box-shadow: none;
-  }
-`
 const Content = styled.div`
   min-height: calc(100vh - 116px) !important;
   ${p => p.contentColor && 'background-color: ' + p.contentColor}
